Fix a typo and improve error handling in VLF_mca/load.py, modify the input of count_mca_intensity() in VLF_mca/mca_intensity_distribution.py
In VLF_mca/load.py, I fixed a typo in the error message printed when the cdf file for mca cannot be opened. Additionally, I improved error handling by adding the Exception to be printed so that we can more easily diagnose the problem.

In VLF_mca/mca_intensity_distribution.py, I modified the input of count_mca_intensity() to take a trange parameter instead of start_date and end_date, for consistency with other functions. The default value of alt_range was also modified to match the range of the available data.