Update intensity distribution plot for electron data in April 1989 and April 1990
The function distribution_plot() in mca_intensity_distribution.py was updated to plot the intensity distribution of electron data in April 1989 and April 1990. The title and save name of the plot were modified to include the date ranges and altitude/MLT ranges. The y-axis label was also modified to include the frequency of the channel. The third dictionary e_dict3 was commented out since it is not used in the function call.

The new version of the function call is distribution_plot(channel=[1, 5], dict_list=[e_dict1, e_dict2]).